pyproject.toml
[project]
name = "frtr"
version = "0.1.0"
description = "FRTR: From Rows to Reasoning – A Multimodal Retrieval Framework for Spreadsheet Understanding"
requires-python = ">=3.10"
dependencies = [
    # Excel processing
    "openpyxl>=3.1.0",
    "pillow>=10.0.0",
    # Vector database
    "chromadb>=0.5.0",
    # BM25 lexical search
    "rank-bm25>=0.2.2",
    # Embeddings - CLIP/SigLIP (local)
    "sentence-transformers>=3.0.0",
    "torch>=2.0.0",
    # Embeddings - OpenAI
    "openai>=1.30.0",
    # AWS Bedrock (optional)
    "boto3>=1.34.0",
    # Code sandbox
    "e2b-code-interpreter>=1.0.0",
    # Utilities
    "numpy>=1.24.0",
    "tqdm>=4.65.0",
    "rich>=13.0.0",
    "click>=8.1.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "httpx[socks]>=0.27.0",
]

[project.scripts]
frtr = "frtr.run_benchmark:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/frtr"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
